pyproject.toml
[project]
name = "invoice-extractor"
version = "0.1.0"
description = "Extract invoice data using Google Gemini + Streamlit"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

authors = [
  {name = "Krishna Kumar", email = "godkrishnasskal@email.com"}
]

dependencies = [
  "streamlit>=1.33.0",
  "google-generativeai>=0.5.2",
  "python-dotenv>=1.0.1",
  "langchain>=0.1.14",
  "PyPDF2>=3.0.1",
  "pandas>=2.2.0"
]

[tool.setuptools]
packages = ["."]
